requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
prophet==1.1.5
xgboost==1.7.6
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
folium==0.14.0
nltk==3.8.1
requests==2.28.2
joblib==1.2.0
weasyprint==59.0
jinja2==3.1.2
pydantic==1.10.7
python-dotenv==1.0.0